chore: move ConverationOptions to it's own file
