+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/conversation/ConverationOptions.kt
@@ -1,0 +1,20 @@
+package com.wire.kalium.logic.data.conversation
+
+data class ConverationOptions(
+    val access: Set<Access> = emptySet(),
+    val accessRole: Set<AccessRole> = emptySet(),
+    val readReceiptsEnabled: Boolean = false,
+    val protocol: Protocol = Protocol.PROTEUS
+) {
+    enum class Protocol {
+        PROTEUS, MLS
+    }
+
+    enum class AccessRole {
+        TEAM_MEMBER, NON_TEAM_MEMBER, GUEST, SERVICE
+    }
+
+    enum class Access {
+        PRIVATE, INVITE, LINK, CODE
+    }
+}
